requirements.txt
requests~=2.32.3
gradio~=5.26.0
pillow~=11.2.1
python-dotenv~=1.1.0
graphviz~=0.20.3
huggingface-hub~=0.30.2
together~=1.5.5
pydantic~=2.11.3
httpx~=0.28.1
pydantic-settings~=2.9.1